syslog-ng-cfg: refactor GlobalConfig
We need two kinds of GlobalConfig: the tests are creating their own
instance and we expect that this instance will be freed by the end of
the tests.

The second kind is used when we receive a GlobalConfig instance from
the C side. We doesn't own this pointer and Rust's strict borrowing
rules also cannot be applied (no guarantee about aliasing XOR mutability,
mutation, etc.). It would be a lie to handle this borrowed raw pointer
as a reference, so I wrapped it into an enum.

This patch also updated the tests and call sites to use the new API.

Signed-off-by: Tibor Benke <ihrwein@gmail.com>